Split Tracing into Kompos and replay support
